[EI-254] [web] feat: Update custom forecast indicator create dialog menu to edit dialog menu
